Completed Todo App
This will be my fallback as I customize and butcher the code to make it my own and continue to learn. I will also take this as an opportunity to get more familiar with Git and Github with branches, and merging, etc. I will branch off new ideas and play with different concepts with this as a framework.